+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGE LOG
 
+
+## 2021.05.14
+- Benchmark the dispatchables in loans pallet (#114)
+
+## 2021.05.12
+- Fix oracle price benchmarking issue (#100)
+
 ## 2021.5.13
 - Add multisig pallet (#112)
 - Add rpc to get price from orml_oracle (#112)
@@ -10,6 +17,9 @@
 - Change LiquidationThreshold from u128 to Ratio (#89)
 - Remove types `OraclePrice`. (#89)
 - Modify types `Price` from `u128` to `FixedU128` (#89)
+
+## 2021.05.09
+- Remove unnecessary dependencies to fix benchmarking (#94)
 
 ## 2021.05.07
 - Add TotalReserves storage (#92)
@@ -30,6 +40,7 @@
 
 ## 2021.04.29
 - Add parallel-dev bin to speed up compilation (#80)
+- Add Benchmarking Infrastructure and Implemented Benchmarking for mint and borrow of pallet-loans (#62)
 
 ## 2021.04.28
 - Modify storage type `SupplyRate` from `u128` to `Rate`. (#82)